feat: refactor common datagrid/table components
