+++ src/components/tables/DecksTable.tsx
@@ -4,7 +4,9 @@ import { useMemo, useRef, useState } from "react";
 
 import {
   DecksTableFilters,
-  DecksTableToolbar,
+  GenericTableToolbar,
+  buildDeckFilterDescription,
+  defaultGridOptions,
   getDecksColumns,
 } from "@/components";
 import {
@@ -15,9 +17,10 @@ import {
   useMatch,
   useTheme,
   useUser,
+  useUserColorRowClasses,
 } from "@/hooks";
 import { Deck, DeckWithStats } from "@/types";
-import { computeDeckStats, getFullColorNames } from "@/utils";
+import { computeDeckStats } from "@/utils";
 
 interface IDecksTableProps {
   customButtons: JSX.Element[];
@@ -37,6 +40,8 @@ export const DecksTable = ({
   const { hasDeckBeenUsed } = useMatch();
   const { mode } = useTheme();
 
+  const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
+  const gridRef = useRef<HTMLDivElement>(null);
   const {
     filterColor,
     setFilterColor,
@@ -48,36 +53,20 @@ export const DecksTable = ({
     setIncludeInactive,
     includeUnranked,
     setIncludeUnranked,
+    resetDeckFilters,
   } = useDecksFilters();
-
-  const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
-  const gridRef = useRef<HTMLDivElement>(null);
 
   const userProfileMap = useMemo(
     () => new Map(allUsers.map((profile) => [profile.id, profile])),
     [allUsers],
   );
 
-  const userColorClasses = useMemo(() => {
-    const styles: Record<string, any> = {};
-    userProfileMap.forEach((user, key) => {
-      const userColor =
-        mode === "light" ? user.lightThemeColor : user.darkThemeColor;
-      styles[`.user-row-${key}`] = {
-        backgroundColor: userColor ? `${userColor}26` : "inherit",
-        border: `1px solid transparent`,
-        "&:hover": {
-          border: (theme: any) => `1px solid ${theme.palette.primary.main}`,
-        },
-      };
-    });
-    return styles;
-  }, [userProfileMap, mode]);
-
-  const deckCategoriesMap = useMemo(
+  const formatsMap = useMemo(
     () => new Map(allFormats.map((format) => [format.id, format])),
     [allFormats],
   );
+
+  const userColorClasses = useUserColorRowClasses(allUsers, mode);
 
   const columns = useMemo(
     () =>
@@ -87,10 +76,10 @@ export const DecksTable = ({
         onDelete,
         currentUserId: userId as string,
         usersMap: userProfileMap,
-        formatsMap: deckCategoriesMap,
+        formatsMap,
         filterFormat: filterFormat,
       }),
-    [userProfileMap, deckCategoriesMap, filterFormat],
+    [userProfileMap, formatsMap, filterFormat],
   );
 
   const filteredDecks: DeckWithStats[] = useMemo(() => {
@@ -117,77 +106,52 @@ export const DecksTable = ({
   ]);
 
   const filterDescription = useMemo(() => {
-    const clauses: string[] = ["showing"];
-    clauses.push(includeInactive ? "all" : "active");
-    if (filterFormat) {
-      clauses.push(`\"${deckCategoriesMap.get(filterFormat)?.displayName}\"`);
-    }
-    clauses.push("decks");
-    if (filterUser) {
-      const userProfile = allUsers.find((profile) => profile.id === filterUser);
-      if (userProfile) {
-        clauses.push(`owned by ${userProfile.displayName}`);
-      }
-    }
-    if (filterColor.length > 0) {
-      clauses.push(
-        filterColor.includes("C") && filterColor.length === 1
-          ? "that are colorless"
-          : `that are ${getFullColorNames(filterColor)}`,
-      );
-    }
-    if (includeUnranked) {
-      clauses.push("while including unranked matches in stats");
-    }
-    return clauses.join(" ");
+    return buildDeckFilterDescription({
+      allFormats,
+      allUsers,
+      filterColor,
+      filterFormat,
+      filterUser,
+      includeInactive,
+      includeUnranked,
+    });
   }, [
-    filterUser,
-    filterFormat,
+    allFormats,
+    allUsers,
     filterColor,
+    filterFormat,
+    filterUser,
     includeInactive,
     includeUnranked,
-    allUsers,
   ]);
 
   return (
     <Stack height="100%" minWidth={740}>
       <DataGrid
+        {...defaultGridOptions}
         ref={gridRef}
-        pagination
-        disableRowSelectionOnClick
         loading={decksLoading}
-        initialState={{
-          density: "compact",
-          pagination: { paginationModel: { pageSize: 15 } },
-        }}
-        pageSizeOptions={[15, 25, 50]}
-        getRowHeight={() => "auto"}
         rows={filteredDecks}
         columns={columns}
         sx={userColorClasses}
-        slots={{
-          toolbar: () => (
-            <DecksTableToolbar
-              customButtons={customButtons}
-              onFilterClick={() => setAnchorEl(gridRef.current)}
-              filterDescription={filterDescription}
-              gridRef={gridRef}
-            />
-          ),
-        }}
-        slotProps={{
-          loadingOverlay: {
-            variant: "skeleton",
-            noRowsVariant: "skeleton",
-          },
-        }}
         getRowClassName={(params: GridRowClassNameParams<Deck>) =>
           `user-row-${params.row.userId}`
         }
+        slots={{
+          toolbar: () => (
+            <GenericTableToolbar
+              customButtons={customButtons}
+              onFilterClick={() => setAnchorEl(gridRef.current)}
+              filterDescription={filterDescription}
+              quickFilterPlaceholder="search decks..."
+            />
+          ),
+        }}
       />
       <DecksTableFilters
         anchorEl={anchorEl}
         onClose={() => setAnchorEl(null)}
+        resetDeckFilters={resetDeckFilters}
         filterColor={filterColor}
         setFilterColor={setFilterColor}
         filterFormat={filterFormat}
