+++ src/components/tables/UsersTable.tsx
@@ -3,8 +3,10 @@ import { DataGrid, GridRowClassNameParams } from "@mui/x-data-grid";
 import { useMemo, useRef, useState } from "react";
 
 import {
+  GenericTableToolbar,
   UsersTableFilters,
-  UsersTableToolbar,
+  buildUserFilterDescription,
+  defaultGridOptions,
   getUsersColumns,
 } from "@/components";
 import {
@@ -12,6 +14,7 @@ import {
   useMatch,
   useTheme,
   useUser,
+  useUserColorRowClasses,
   useUsersFilters,
 } from "@/hooks";
 import { User, UserWithStats } from "@/types";
@@ -27,6 +30,10 @@ export const UsersTable = ({ customButtons }: IUsersTableProps) => {
   const { allFormats } = useFormat();
   const { mode } = useTheme();
 
+  const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
+  const gridRef = useRef<HTMLDivElement>(null);
+  const userColorClasses = useUserColorRowClasses(allUsers, mode);
+
   const {
     filterFormat,
     setFilterFormat,
@@ -34,26 +41,8 @@ export const UsersTable = ({ customButtons }: IUsersTableProps) => {
     setIncludeUnranked,
     activeRecentOnly,
     setActiveRecentOnly,
+    resetUserFilters,
   } = useUsersFilters();
-
-  const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
-  const gridRef = useRef<HTMLDivElement>(null);
-
-  const userColorClasses = useMemo(() => {
-    const styles: Record<string, any> = {};
-    allUsers.forEach((user) => {
-      const userColor =
-        mode === "light" ? user.lightThemeColor : user.darkThemeColor;
-      styles[`.user-row-${user.id}`] = {
-        backgroundColor: userColor ? `${userColor}26` : "inherit",
-        border: `1px solid transparent`,
-        "&:hover": {
-          border: (theme: any) => `1px solid ${theme.palette.primary.main}`,
-        },
-      };
-    });
-    return styles;
-  }, [allUsers, mode]);
 
   const columns = useMemo(() => getUsersColumns(), []);
 
@@ -77,61 +66,41 @@ export const UsersTable = ({ customButtons }: IUsersTableProps) => {
   ]);
 
   const filterDescription = useMemo(() => {
-    const formatName = allFormats.find(
-      (f) => f.id === filterFormat,
-    )?.displayName;
-    const clauses = ["showing stats for users"];
-    if (activeRecentOnly) {
-      clauses.push("active in the last 30 days");
-    }
-    if (formatName) {
-      clauses.push(`for the format \"${formatName}\"`);
-    }
-    if (!formatName && includeUnranked) {
-      clauses.push("(including unranked matches)");
-    }
-    return clauses.join(" ");
-  }, [filterFormat, includeUnranked, allFormats]);
+    return buildUserFilterDescription({
+      allFormats,
+      activeRecentOnly,
+      includeUnranked,
+      filterFormat,
+    });
+  }, [filterFormat, includeUnranked, allFormats, activeRecentOnly]);
 
   return (
     <Stack height="100%" minWidth={740}>
       <DataGrid
+        {...defaultGridOptions}
         ref={gridRef}
-        pagination
-        disableRowSelectionOnClick
         loading={usersLoading}
-        initialState={{
-          density: "compact",
-          pagination: { paginationModel: { pageSize: 15 } },
-        }}
-        pageSizeOptions={[15, 25, 50]}
-        getRowHeight={() => "auto"}
         rows={usersLoading ? [] : filteredUsers}
         columns={columns}
         sx={userColorClasses}
-        slots={{
-          toolbar: () => (
-            <UsersTableToolbar
-              customButtons={customButtons}
-              onFilterClick={() => setAnchorEl(gridRef.current)}
-              filterDescription={filterDescription}
-              gridRef={gridRef}
-            />
-          ),
-        }}
-        slotProps={{
-          loadingOverlay: {
-            variant: "skeleton",
-            noRowsVariant: "skeleton",
-          },
-        }}
         getRowClassName={(params: GridRowClassNameParams<User>) =>
           `user-row-${params.row.id}`
         }
+        slots={{
+          toolbar: () => (
+            <GenericTableToolbar
+              customButtons={customButtons}
+              onFilterClick={() => setAnchorEl(gridRef.current)}
+              filterDescription={filterDescription}
+              quickFilterPlaceholder="search users..."
+            />
+          ),
+        }}
       />
       <UsersTableFilters
         anchorEl={anchorEl}
         onClose={() => setAnchorEl(null)}
+        resetUserFilters={resetUserFilters}
         filterFormat={filterFormat}
         setFilterFormat={setFilterFormat}
         includeUnranked={includeUnranked}
