+++ src/components/tables/datagrid/columns/index.ts
@@ -1,0 +1,3 @@
+export * from "./getDeckColumns";
+export * from "./getMatchColumns";
+export * from "./getUserColumns";
