--- src/components/tables/decks/index.ts
@@ -1,4 +1,0 @@
-export * from "./DecksTable";
-export * from "./DecksTableColumns";
-export * from "./DecksTableFilters";
-export * from "./DecksTableToolbar";
