+++ src/components/tables/index.ts
@@ -1,5 +1,4 @@
-export * from "./DeckColorCell";
-export * from "./EnhancedTableHeader";
-export * from "./decks";
-export * from "./matches";
-export * from "./users";
+export * from "./DecksTable";
+export * from "./MatchesTable";
+export * from "./UsersTable";
+export * from "./datagrid";
