--- src/components/tables/matches/index.ts
@@ -1,4 +1,0 @@
-export * from "./MatchesTable";
-export * from "./MatchesTableColumns";
-export * from "./MatchesTableFilters";
-export * from "./MatchesTableToolbar";
