--- src/components/tables/users/index.ts
@@ -1,4 +1,0 @@
-export * from "./UsersTable";
-export * from "./UsersTableColumns";
-export * from "./UsersTableFilters";
-export * from "./UsersTableToolbar";
