+++ src/hooks/useMatchFilters.ts
@@ -1,20 +1,26 @@
 import { usePersistentState } from "@/hooks";
 
 export const useMatchesFilters = () => {
-  const [filterFormat, setFilterFormat] = usePersistentState<string>(
-    "matches_filter_format",
-    "",
-  );
-
   const [filterDecks, setFilterDecks] = usePersistentState<string[]>(
     "matches_filter_decks",
     [],
+  );
+
+  const [filterFormat, setFilterFormat] = usePersistentState<string>(
+    "matches_filter_format",
+    "",
   );
 
   const [filterUsers, setFilterUsers] = usePersistentState<string[]>(
     "matches_filter_users",
     [],
   );
+
+  const resetMatchFilters = () => {
+    setFilterDecks([]);
+    setFilterFormat("");
+    setFilterUsers([]);
+  }
 
   return {
     filterFormat,
@@ -23,5 +29,6 @@ export const useMatchesFilters = () => {
     setFilterDecks,
     filterUsers,
     setFilterUsers,
+    resetMatchFilters,
   };
 };
